Step screen complete

+++ app/src/main/java/com/bill/android/baking101/models/Ingredient.java
@@ -5,21 +5,21 @@ import android.os.Parcelable;
 
 public class Ingredient implements Parcelable {
 
-    private int mQuantity;
+    private double mQuantity;
     private String mMeasure;
     private String mName;
 
-    public Ingredient(int quantity, String measure, String name) {
+    public Ingredient(double quantity, String measure, String name) {
         mQuantity = quantity;
         mMeasure = measure;
         mName = name;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return mQuantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.mQuantity = quantity;
     }
 
@@ -49,7 +49,7 @@ public class Ingredient implements Parcelable {
     }
 
     protected Ingredient(Parcel in) {
-        mQuantity = in.readInt();
+        mQuantity = in.readDouble();
         mMeasure = in.readString();
         mName = in.readString();
     }
@@ -73,7 +73,7 @@ public class Ingredient implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel parcel, int i) {
-        parcel.writeInt(mQuantity);
+        parcel.writeDouble(mQuantity);
         parcel.writeString(mMeasure);
         parcel.writeString(mName);
     }
